Alterações de melhoria e organização do código, também adiciona o status da tarefa nos detalhes.
