+++ src/style.css
@@ -192,6 +192,7 @@
     text-align: left;
     color: black;
     font-size: 18px;
+    cursor: pointer;
 }
 
 #task-details{
