+++ src/utils/storage.ts
@@ -1,20 +1,6 @@
-import ITaskModel from "../models/Task.js";
+import { Task } from "../models/Task";
 
 export let tasks: Task[] = [];
-export class Task implements ITaskModel {
-    name: string;
-    description: string;
-    date: Date;
-    id: number;
-    status: boolean;
-    constructor(name:string, description: string, date: Date, id: number, status: boolean) {
-        this.name = name;
-        this.description = description;
-        this.date = date;
-        this.id = id;
-        this.status = status;
-    }
-}
 
 export function saveTasksToLocalStorage() {
     localStorage.setItem("tasks", JSON.stringify(tasks));
